fx_dashboard/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
lseg-data>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
websockets>=12.0
